+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @0xsequence/hooks
 
+## 5.4.8
+
+### Patch Changes
+
+- sequence.js dependency upgrades
+
+## 5.4.7
+
+### Patch Changes
+
+- sequence.js upgrades
+
+## 5.4.6
+
+### Patch Changes
+
+- Dependency updates
+
+## 5.4.5
+
+### Patch Changes
+
+- Fix for free items checkout, sequence dep upgrades
+
+## 5.4.4
+
+### Patch Changes
+
+- Sequence logo update
+
 ## 5.4.3
 
 ### Patch Changes